requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
easyocr>=1.7.0
opencv-python>=4.8.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
transformers>=4.35.0
sentence-transformers>=2.2.0
torch>=2.1.0
chromadb>=0.4.18
google-generativeai>=0.3.0